v0.3.0

+++ crates/diman_unit_system/src/dimension_math.rs
@@ -56,7 +56,7 @@ impl MulDiv for DimensionsAndFactor {
                     .into_iter()
                     .map(|entry| DimensionEntry {
                         ident: entry.ident,
-                        value: entry.value * pow as i32,
+                        value: entry.value * pow,
                     })
                     .collect(),
             },
